+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # React Native Module 26.0.0 Changelog
 
+## Version 26.8.0 - July 13, 2026
+
+Minor release that updates the Android SDK to 20.9.0 and the iOS SDK to 20.10.0; adds improved prioritization control for embedded displays, and fixes a bug causing messages to fail to render in the embedded message view.
+
+### Changes
+- Updated iOS SDK to [20.10.0](https://github.com/urbanairship/ios-library/releases/tag/20.10.0)
+- Updated Android SDK to [20.9.0](https://github.com/urbanairship/android-library/releases/tag/20.9.0)
+- Fixed native (Scenes) messages failing to render in the embedded message view
+- Added a `selection` parameter to `AirshipEmbeddedView` to pick which pending instance displays, by priority or a specific instance ID
+
 ## Version 26.7.0 - June 25, 2026
 
 Minor release that updates the iOS SDK to 20.8.1 and the Android SDK to 20.8.1.